requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
python-telegram-bot==13.15
pytz==2024.1
setuptools==70.0.0
six==1.16.0
tornado==6.1
tzlocal==5.2
emoji==1.7.0
clarifai-grpc==7.11.0
dnspython==1.16.0
pymongo==3.10.1